In merge dialog also Start-ID should be searched
In some school events, when renting 200-500 si-cards, we give to every
runner a random si-card, but a specific start number. At the read out
place, when then fill in the start number the runner is wearing. So it
would be nice if in Geco in the merge dialog you could also search after
the start number